Add RemoteTileClient
