+++ localtileserver/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa: F401
 from localtileserver._version import __version__
+from localtileserver.client import RemoteTileClient
 from localtileserver.report import Report
 from localtileserver.server import TileClient, get_or_create_tile_client
 from localtileserver.utilities import get_cache_dir, make_vsi, purge_cache
