+++ localtileserver/server.py
@@ -1,17 +1,12 @@
 import logging
 import pathlib
 import threading
-from typing import List, Union
+from typing import Union
 
 import requests
 from werkzeug.serving import make_server
 
-from localtileserver.palettes import palette_valid_or_raise
-from localtileserver.utilities import (
-    add_query_parameters,
-    get_clean_filename,
-    save_file_from_request,
-)
+from localtileserver.client import BaseTileClient, RemoteTileClient
 
 _LIVE_SERVERS = {}
 
@@ -117,7 +112,7 @@ def shutdown_server(key: int, force: bool = False):
         logger.error(f"Server for key ({key}) not found.")
 
 
-class TileClient:
+class TileClient(BaseTileClient):
     """Serve tiles from a local raster file in a background thread.
 
     Parameters
@@ -144,17 +139,13 @@ class TileClient:
         threaded: bool = True,
         processes: int = 1,
     ):
-        self._filename = get_clean_filename(filename)
+        super().__init__(filename)
         self._key = launch_server(port, debug, threaded=threaded, processes=processes)
         # Store actual port just in case
         self._port = _LIVE_SERVERS[self._key].srv.port
 
     def __del__(self):
         self.shutdown()
-
-    @property
-    def filename(self):
-        return self._filename
 
     @property
     def server(self):
@@ -168,145 +159,15 @@ class TileClient:
         return self.server.port
 
     @property
+    def host(self):
+        return self.server.host
+
+    @property
     def base_url(self):
-        return f"http://{self.server.host}:{self.port}"
+        return f"http://{self.host}:{self.port}"
 
     def shutdown(self, force: bool = False):
         shutdown_server(self._key, force=force)
-
-    def _produce_url(self, base: str):
-        return add_query_parameters(base, {"filename": self._filename})
-
-    def create_url(self, path: str):
-        return self._produce_url(f"{self.base_url}/{path.lstrip('/')}")
-
-    def get_tile_url(
-        self,
-        projection: str = "EPSG:3857",
-        band: Union[int, List[int]] = None,
-        palette: Union[str, List[str]] = None,
-        vmin: Union[Union[float, int], List[Union[float, int]]] = None,
-        vmax: Union[Union[float, int], List[Union[float, int]]] = None,
-        nodata: Union[Union[float, int], List[Union[float, int]]] = None,
-    ):
-        """
-
-        Parameters
-        ----------
-        projection : str
-            The Proj projection to use for the tile layer. Default is `EPSG:3857`.
-        band : int
-            The band of the source raster to use (default in None to show RGB if
-            available). Band indexing starts at 1. This can also be a list of
-            integers to set which 3 bands to use for RGB.
-        palette : str
-            The name of the color palette from `palettable` or colormap from
-            matplotlib to use when plotting a single band. Default is greyscale.
-        vmin : float
-            The minimum value to use when colormapping the palette when plotting
-            a single band.
-        vmax : float
-            The maximized value to use when colormapping the palette when plotting
-            a single band.
-        nodata : float
-            The value from the band to use to interpret as not valid data.
-
-        """
-        # First handle query parameters to check for errors
-        params = {}
-        if band is not None:
-            params["band"] = band
-        if palette is not None:
-            # make sure palette is valid
-            palette_valid_or_raise(palette)
-            params["palette"] = palette
-        if vmin is not None:
-            params["min"] = vmin
-        if vmax is not None:
-            params["max"] = vmax
-        if nodata is not None:
-            params["nodata"] = nodata
-        if projection is not None:
-            params["projection"] = projection
-        return add_query_parameters(self.create_url("tiles/{z}/{x}/{y}.png"), params)
-
-    def extract_roi(
-        self,
-        left: float,
-        right: float,
-        bottom: float,
-        top: float,
-        units: str = "EPSG:4326",
-        encoding: str = "TILED",
-        output_path: pathlib.Path = None,
-    ):
-        """Extract ROI in world coordinates."""
-        path = f"/region/world/{left}/{right}/{bottom}/{top}/region.tif?units={units}&encoding={encoding}"
-        r = requests.get(self.create_url(path))
-        r.raise_for_status()
-        return save_file_from_request(r, output_path)
-
-    def extract_roi_pixel(
-        self,
-        left: int,
-        right: int,
-        bottom: int,
-        top: int,
-        encoding: str = "TILED",
-        output_path: pathlib.Path = None,
-    ):
-        """Extract ROI in world coordinates."""
-        path = f"/region/pixel/{left}/{right}/{bottom}/{top}/region.tif?encoding={encoding}"
-        r = requests.get(self.create_url(path))
-        r.raise_for_status()
-        return save_file_from_request(r, output_path)
-
-    def metadata(self):
-        r = requests.get(self.create_url("/metadata"))
-        r.raise_for_status()
-        return r.json()
-
-    def bounds(self, projection: str = "EPSG:4326"):
-        """Get bounds in form of (ymin, ymax, xmin, xmax)."""
-        r = requests.get(self.create_url(f"/bounds?projection={projection}"))
-        r.raise_for_status()
-        bounds = r.json()
-        return (bounds["ymin"], bounds["ymax"], bounds["xmin"], bounds["xmax"])
-
-    def center(self, projection: str = "EPSG:4326"):
-        """Get center in the form of (y <lat>, x <lon>)."""
-        bounds = self.bounds(projection=projection)
-        return (
-            (bounds[1] - bounds[0]) / 2 + bounds[0],
-            (bounds[3] - bounds[2]) / 2 + bounds[2],
-        )
-
-    def thumbnail(
-        self,
-        band: Union[int, List[int]] = None,
-        palette: Union[str, List[str]] = None,
-        vmin: Union[Union[float, int], List[Union[float, int]]] = None,
-        vmax: Union[Union[float, int], List[Union[float, int]]] = None,
-        nodata: Union[Union[float, int], List[Union[float, int]]] = None,
-        output_path: pathlib.Path = None,
-    ):
-        params = {}
-        if band is not None:
-            params["band"] = band
-        if palette is not None:
-            # make sure palette is valid
-            palette_valid_or_raise(palette)
-            params["palette"] = palette
-        if vmin is not None:
-            params["min"] = vmin
-        if vmax is not None:
-            params["max"] = vmax
-        if nodata is not None:
-            params["nodata"] = nodata
-        url = add_query_parameters(self.create_url("thumbnail"), params)
-        r = requests.get(url)
-        r.raise_for_status()
-        return save_file_from_request(r, output_path)
 
 
 def get_or_create_tile_client(
@@ -323,6 +184,8 @@ def get_or_create_tile_client(
     default is for all TileClient's to share a single server.
 
     """
+    if isinstance(source, RemoteTileClient):
+        return source, False
     _internally_created = False
     # Launch tile server if file path is given
     if not isinstance(source, TileClient):
